+++ src/runtime.rs
@@ -4,7 +4,7 @@
 //  Created:
 //    10 Dec 2024, 17:11:17
 //  Last edited:
-//    13 Jan 2025, 14:28:17
+//    13 Jan 2025, 17:02:30
 //  Auto updated?
 //    Yes
 //
@@ -23,7 +23,8 @@ use crate::actions::Action;
 use crate::actors::{Agent, Synchronizer};
 use crate::agreements::Agreement;
 use crate::auxillary::{Authored, Identifiable};
-use crate::collections::{InfallibleMap, Map, SetMut};
+use crate::collections::map::{InfallibleMap, Map, MapAsync, MapSync};
+use crate::collections::set::SetSync;
 use crate::messages::{Message, MessageSet};
 use crate::times::Times;
 
@@ -193,19 +194,22 @@ impl<T, A, S, E> View<T, A, S, E> {
 /// Defines the toplevel [`Runtime`], which brings the ontology together.
 #[pointer_impls]
 pub trait Runtime {
+    /// Defines the type of identifiers for agents / synchronizers.
+    type AgentId;
+
     /// Defines the type of messages in the runtime.
     type Message: Message;
     /// Defines the type of actions in the runtime.
     type Action: Action<Message = Self::Message, Timestamp = <Self::Times as Times>::Timestamp>;
 
     /// Defines the set of synchronized times.
-    type Times: SetMut<<Self::Times as Times>::Timestamp> + Times;
+    type Times: SetSync<<Self::Times as Times>::Timestamp> + Times;
     /// Defines the set of synchronized agreements.
-    type Agreements: SetMut<Agreement<Self::Message, <Self::Times as Times>::Timestamp>>;
+    type Agreements: MapSync<Agreement<Self::Message, <Self::Times as Times>::Timestamp>>;
     /// Defines the set of statements.
-    type Statements: SetMut<Self::Message>;
+    type Statements: MapAsync<Self::AgentId, Self::Message>;
     /// Defines the set of enacted actions.
-    type Enactments: SetMut<Self::Action>;
+    type Enactments: MapAsync<Self::AgentId, Self::Action>;
 
     /// Any errors thrown by the runtime.
     type Error: error::Error;
@@ -220,7 +224,7 @@ pub trait Runtime {
     ///
     /// # Errors
     /// This function errors whenever anything in the set goes wrong.
-    fn run<A>(&mut self, agents: impl IntoIterator<Item = A>, synchronizer: impl Synchronizer) -> Result<(), Self::Error>
+    fn run<A>(&mut self, agents: impl IntoIterator<Item = A>, synchronizer: impl Synchronizer<Id = Self::AgentId>) -> Result<(), Self::Error>
     where
-        A: Agent;
+        A: Agent<Id = Self::AgentId>;
 }
